Add dsub version and restore full setup.py
Dsub now has a version field available in each of the following;
dsub.__init__, dsub.providers.base, and setup.py. This reference
allows the use of version semantics defined in PEP-396

    import dsub
    print(dsub.__version__)

dsub also has a reference to version inside dsub.providers.base
allowing the google provider to attach version metadata to pipelines
requests. Of critical importance is the fact that none of the
references to version depend on the ability to import a specific dsub
module, this prevents cyclic dependency (dsub -> providers -> dsub)
and it prevents common bugs in distutils (rule of thumb: never import
the target module from setup.py).

PiperOrigin-RevId: 169743093