dsub with --skip and --tasks
Previously, dsub supported both --tasks (run multiple programs as part of one job)
and --skip (do not execute if the output already exists), but only separately.

With this change, the two options can now be used together. This is compelling
since when running many tasks, it is even more useful to be able to skip the ones
whose output is already present (perhaps because of an earlier run where some tasks
failed and some did not).

PiperOrigin-RevId: 181680285