Add script name to error message in local provider
dsub generates three logs: the provider log, user script stdout,
and user script stderr.

When the provider script encounters an error, it logs a line number.
But this can be confusing to the user because they don't know which
file this line number comes from.

To clear things up, this CL adds the script name to the log when
there's an error with the local provider.

PiperOrigin-RevId: 211861596